requirements.txt
scikit-image
scipy
opencv-python
numpy
Pillow
natsort
send2trash
lxml
requests
ipython
ffmpeg-python